requirement.txt
Django
jproperties
logging
djangorestframework